Step 3 - Scripting
Added most of the scripts for the objects in apple picker. So far, I've
gotten the AppleTree to change directions and drop apples. I haven't
incorporated the basket and trigger collisions into the game just yet.

It was somewhat difficult to follow the code in the text book due to how
far he spaces it out. While I do like how it's shown iteration by
iteration, I don't like that it doesn't show the final product of the
script at the end.

Due to this, I accidentally had some functions in apple that belonged to
apple tree causing a half hour of confusion.