Step 7 - High Score & Final stetch
I added code for the game to reset the scene/level if the basket count
hits zero. Also added a script for High score which allows it to act as
a static variable which is updated and saved accordingly throughout game
sessions.

Although I followed the textbook to the letter, the scores on the screen
will not change when I pick up an apple and I can't find out why. I've
gone over the relevant code a few times now to no avail.

Maybe the text object I added isn't the same thing as the GUIText
referenced in the book?